Refactor flight track widget and share
